vehicle-detect/requirements.txt
pika
python-dotenv
opencv-python-headless
numpy
requests
matplotlib
Pillow
torch
torchvision
tqdm
pyyaml
pandas
seaborn
scipy
